requirements.txt
astroid==2.4.2
click==7.1.2
colorama==0.4.4
DateTime==4.3
fastapi==0.63.0
h11==0.12.0
isort==5.7.0
lazy-object-proxy==1.4.3
mccabe==0.6.1
numpy==1.20.0
pandas==1.2.1
psycopg2==2.8.6
psycopg2-binary==2.8.6
pydantic==1.7.3
pylint==2.6.0
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2021.1
six==1.15.0
starlette==0.13.6
toml==0.10.2
uvicorn==0.13.3
wrapt==1.12.1
zope.interface==5.2.0
